feat: wire up add coupon form and wire up delete coupon button

+++ src/BeanVault.Services/BeanVault.Services.CouponService/BeanVault.Services.CouponService.Core/Models/Coupon.cs
@@ -5,5 +5,5 @@ public class Coupon
   public string Id { get; set; } = string.Empty;
   public string CouponCode { get; set; } = string.Empty;
   public double DiscountAmount { get; set; }
-  public int MinAmount { get; set; }
+  public double MinAmount { get; set; }
 }
